Rename sat_solver.py -> solver.py
