+++ satsolver/solver.py
